Changed a lot. DELETE CUSTOMERDATABASE FILE BEFORE RUNNING. Changed the account ID to string for checking and savings. my idea is that I'm going to have standard account ids. So the first account gets ID 1. and if they have a checking account it gets the id 1-00 if they have a simple savings that account gets 1-01, savings cd gets 1-02. loan accounts will have 03,04-05 etc, But that customer will have a consistent account Id. I changed it to discard those two checking accounts with all null values. The base data is still there, but they have no checking account.
NOW FOR THE GUI

I added the interfaces to manage loan accounts. I made all the changes above,
because I didn't want to have to search two data sets to check for existing
account ID for the savings, checkings, and cd accounts. So now I can just generate
an ID and use it for all accounts under a single customer. So for the Add checking
interface, All they should need is the starting balance, if it's a gold account,
and if they want backup savings.