Fixed all enterKey events to fire on a key release instead of key press, which prevents it from looping through two interfaces.
Added a serialize to file
after each change that changes or modifies customer data. So it is written to the
File and if the program crashes, the file should still be fine.

Made a Logout button that logs out the Manager Or Teller account if before the
5 minute cutoff.

Lastly, Added validation to prevent a customer from withdrawaling more than
Two times a day according to Project Instructions.